Remove file explorer
