--- src/browser/features/RightSidebar/FileViewer/index.ts
@@ -1,7 +1,0 @@
-/**
- * FileViewer - Components for viewing files in the right sidebar.
- */
-
-export { FileViewerTab } from "./FileViewerTab";
-export { TextFileViewer } from "./TextFileViewer";
-export { ImageFileViewer } from "./ImageFileViewer";
