+++ src/browser/features/RightSidebar/RightSidebar.stories.tsx
@@ -390,172 +390,6 @@ export const ReviewTab: Story = {
 };
 
 /**
- * Explorer tab showing workspace file tree with folders and files
- */
-export const ExplorerTab: Story = {
-  render: () => (
-    <RightSidebarStoryShell
-      setup={() => {
-        localStorage.setItem(RIGHT_SIDEBAR_TAB_KEY, JSON.stringify("explorer"));
-        localStorage.setItem(RIGHT_SIDEBAR_WIDTH_KEY, "350");
-        localStorage.removeItem(getRightSidebarLayoutKey("ws-explorer"));
-        localStorage.removeItem(`explorer:entries:ws-explorer`);
-        localStorage.removeItem(`explorer:expanded:ws-explorer`);
-
-        const client = setupSimpleChatStory({
-          workspaceId: "ws-explorer",
-          workspaceName: "feature/files",
-          projectName: "my-app",
-          messages: [
-            createUserMessage("msg-1", "Show me the project structure", { historySequence: 1 }),
-            createAssistantMessage("msg-2", "Here is the project structure.", {
-              historySequence: 2,
-            }),
-          ],
-        });
-        expandRightSidebar();
-        return client;
-      }}
-    >
-      <RightSidebarStoryContent workspaceId="ws-explorer" />
-    </RightSidebarStoryShell>
-  ),
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-
-    // Wait for explorer tab to be available and click it
-    const explorerTab = await canvas.findByRole("tab", { name: /^explorer/i });
-    await userEvent.click(explorerTab);
-
-    // Wait for file tree to load (mock returns src, tests, node_modules, etc.)
-    await waitFor(() => {
-      canvas.getByText("src");
-      canvas.getByText("package.json");
-    });
-
-    // Verify ignored folder is shown with reduced opacity (node_modules)
-    await waitFor(() => {
-      canvas.getByText("node_modules");
-    });
-  },
-};
-
-/**
- * Explorer tab with expanded directory showing Collapse All button
- */
-export const ExplorerTabExpanded: Story = {
-  render: () => (
-    <RightSidebarStoryShell
-      setup={() => {
-        localStorage.setItem(RIGHT_SIDEBAR_TAB_KEY, JSON.stringify("explorer"));
-        localStorage.setItem(RIGHT_SIDEBAR_WIDTH_KEY, "350");
-        localStorage.removeItem(getRightSidebarLayoutKey("ws-explorer-expanded"));
-        localStorage.removeItem(`explorer:entries:ws-explorer-expanded`);
-        localStorage.removeItem(`explorer:expanded:ws-explorer-expanded`);
-
-        const client = setupSimpleChatStory({
-          workspaceId: "ws-explorer-expanded",
-          workspaceName: "feature/files",
-          projectName: "my-app",
-          messages: [
-            createUserMessage("msg-1", "Show me the project structure", { historySequence: 1 }),
-            createAssistantMessage("msg-2", "Here is the project structure.", {
-              historySequence: 2,
-            }),
-          ],
-        });
-        expandRightSidebar();
-        return client;
-      }}
-    >
-      <RightSidebarStoryContent workspaceId="ws-explorer-expanded" />
-    </RightSidebarStoryShell>
-  ),
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-
-    // Wait for explorer tab and click it
-    const explorerTab = await canvas.findByRole("tab", { name: /^explorer/i });
-    await userEvent.click(explorerTab);
-
-    // Wait for file tree to load
-    await waitFor(() => {
-      canvas.getByText("src");
-    });
-
-    // Click on src folder to expand it
-    const srcFolder = canvas.getByText("src");
-    await userEvent.click(srcFolder);
-
-    // Wait for src contents to load and collapse all button to appear
-    await waitFor(() => {
-      canvas.getByText("App.tsx");
-      canvas.getByText("components");
-    });
-
-    // Verify collapse all button is visible (tooltip text)
-    await waitFor(() => {
-      canvas.getByRole("button", { name: /collapse all/i });
-    });
-
-    // Blur to get clean screenshot
-    blurActiveElement();
-  },
-};
-
-/**
- * Explorer tab with selected item showing blue background
- */
-export const ExplorerTabSelected: Story = {
-  render: () => (
-    <RightSidebarStoryShell
-      setup={() => {
-        localStorage.setItem(RIGHT_SIDEBAR_TAB_KEY, JSON.stringify("explorer"));
-        localStorage.setItem(RIGHT_SIDEBAR_WIDTH_KEY, "350");
-        localStorage.removeItem(getRightSidebarLayoutKey("ws-explorer-selected"));
-        localStorage.removeItem(`explorer:entries:ws-explorer-selected`);
-        localStorage.removeItem(`explorer:expanded:ws-explorer-selected`);
-
-        const client = setupSimpleChatStory({
-          workspaceId: "ws-explorer-selected",
-          workspaceName: "feature/files",
-          projectName: "my-app",
-          messages: [
-            createUserMessage("msg-1", "Show me the project structure", { historySequence: 1 }),
-            createAssistantMessage("msg-2", "Here is the project structure.", {
-              historySequence: 2,
-            }),
-          ],
-        });
-        expandRightSidebar();
-        return client;
-      }}
-    >
-      <RightSidebarStoryContent workspaceId="ws-explorer-selected" />
-    </RightSidebarStoryShell>
-  ),
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-
-    // Wait for explorer tab and click it
-    const explorerTab = await canvas.findByRole("tab", { name: /^explorer/i });
-    await userEvent.click(explorerTab);
-
-    // Wait for file tree to load
-    await waitFor(() => {
-      canvas.getByText("src");
-    });
-
-    // Click on src folder to select it (will have focus/selected blue background)
-    // Using a folder instead of a file to avoid opening a file viewer tab
-    const srcFolder = canvas.getByText("src");
-    await userEvent.click(srcFolder);
-
-    // Don't blur - keep the item selected/focused for the screenshot
-  },
-};
-
-/**
  * Stats tab (Timing sub-tab) when idle (no timing data) - shows placeholder message
  */
 export const StatsTabIdle: Story = {
@@ -1350,69 +1184,6 @@ export const ReviewTabWithUntrackedFiles: Story = {
     await new Promise((r) => requestAnimationFrame(() => requestAnimationFrame(r)));
 
     // Blur focus for clean screenshot
-    blurActiveElement();
-  },
-};
-
-/**
- * Many tabs at a narrow sidebar width should wrap into multiple rows.
- */
-export const ManyTabsWrap: Story = {
-  render: () => (
-    <RightSidebarStoryShell
-      setup={() => {
-        const workspaceId = "ws-many-tabs";
-        updatePersistedState(RIGHT_SIDEBAR_WIDTH_KEY, 280);
-        updatePersistedState(getRightSidebarLayoutKey(workspaceId), {
-          version: 1,
-          nextId: 2,
-          focusedTabsetId: "tabset-1",
-          root: {
-            type: "tabset",
-            id: "tabset-1",
-            tabs: [
-              "costs",
-              "review",
-              "explorer",
-              ...Array.from(
-                { length: 12 },
-                (_v, i) => `file:src/components/ThisIsAReallyLongFileName${i + 1}.tsx`
-              ),
-            ],
-            activeTab: "review",
-          },
-        });
-
-        const client = setupSimpleChatStory({
-          workspaceId,
-          workspaceName: "feature/many-tabs",
-          projectName: "my-app",
-          messages: [
-            createUserMessage("msg-1", "Open a lot of tabs", { historySequence: 1 }),
-            createAssistantMessage("msg-2", "Ok.", { historySequence: 2 }),
-          ],
-        });
-        expandRightSidebar();
-        return client;
-      }}
-    >
-      <RightSidebarStoryContent workspaceId="ws-many-tabs" />
-    </RightSidebarStoryShell>
-  ),
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-
-    await canvas.findByRole("tablist", { name: /sidebar views/i }, { timeout: 10_000 });
-
-    // Re-query tablist inside waitFor so retries always use a fresh DOM ref
-    // (the captured ref from findByRole could go stale if the component remounts).
-    await waitFor(async () => {
-      const tablist = canvas.getByRole("tablist", { name: /sidebar views/i });
-      const tabs = within(tablist).getAllByRole("tab");
-      const rowTops = new Set(tabs.map((tab) => Math.round(tab.getBoundingClientRect().top)));
-      await expect(rowTops.size).toBeGreaterThan(1);
-    });
-
     blurActiveElement();
   },
 };
