+++ src/browser/features/RightSidebar/Tabs/TabLabels.tsx
@@ -8,17 +8,8 @@
  */
 
 import React from "react";
-import {
-  BugPlay,
-  ExternalLink,
-  FolderTree,
-  Monitor,
-  Globe,
-  Terminal as TerminalIcon,
-  X,
-} from "lucide-react";
+import { BugPlay, ExternalLink, Monitor, Globe, Terminal as TerminalIcon, X } from "lucide-react";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/browser/components/Tooltip/Tooltip";
-import { FileIcon } from "@/browser/components/FileIcon/FileIcon";
 import { type ReviewStats } from "./registry";
 import { formatKeybind, KEYBINDS } from "@/browser/utils/ui/keybinds";
 import { cn } from "@/common/lib/utils";
@@ -87,14 +78,6 @@ export const ReviewTabLabel: React.FC<ReviewTabLabelProps> = ({ reviewStats }) =
   </>
 );
 
-/** Explorer tab label with folder tree icon */
-export const ExplorerTabLabel: React.FC = () => (
-  <span className="inline-flex items-center gap-1">
-    <FolderTree className="h-3 w-3 shrink-0" />
-    Explorer
-  </span>
-);
-
 /** Desktop tab label with monitor icon */
 export const DesktopTabLabel: React.FC = () => (
   <span className="inline-flex items-center gap-1">
@@ -122,44 +105,6 @@ export const DebugTabLabel: React.FC = () => (
 export function OutputTabLabel() {
   return <>Output</>;
 }
-
-interface FileTabLabelProps {
-  /** File path (relative to workspace) */
-  filePath: string;
-  /** Callback when close button is clicked */
-  onClose: () => void;
-}
-
-/** File tab label with file icon, filename, and close button */
-export const FileTabLabel: React.FC<FileTabLabelProps> = ({ filePath, onClose }) => {
-  // Extract just the filename for display
-  const fileName = filePath.split("/").pop() ?? filePath;
-
-  return (
-    <span className="inline-flex items-center gap-1">
-      <FileIcon fileName={fileName} style={{ fontSize: 14 }} className="h-3.5 w-3.5 shrink-0" />
-      <span className="max-w-[120px] truncate" title={filePath}>
-        {fileName}
-      </span>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <button
-            type="button"
-            className="text-muted hover:text-destructive -my-0.5 rounded p-0.5 transition-colors"
-            onClick={(e) => {
-              e.stopPropagation();
-              onClose();
-            }}
-            aria-label="Close file"
-          >
-            <X className="h-3 w-3" />
-          </button>
-        </TooltipTrigger>
-        <TooltipContent side="bottom">Close ({formatKeybind(KEYBINDS.CLOSE_TAB)})</TooltipContent>
-      </Tooltip>
-    </span>
-  );
-};
 
 interface TerminalTabLabelProps {
   /** Dynamic title from OSC sequences, if available */
