+++ src/browser/features/RightSidebar/Tabs/index.ts
@@ -9,7 +9,6 @@
 export {
   TAB_CONFIGS,
   TERMINAL_TAB_CONFIG,
-  FILE_TAB_CONFIG,
   getTabConfig,
   getTabName,
   getTabContentClassName,
@@ -20,11 +19,4 @@ export {
   type ReviewStats,
 } from "./registry";
 
-export {
-  StatsTabLabel,
-  ExplorerTabLabel,
-  OutputTabLabel,
-  FileTabLabel,
-  ReviewTabLabel,
-  TerminalTabLabel,
-} from "./TabLabels";
+export { StatsTabLabel, OutputTabLabel, ReviewTabLabel, TerminalTabLabel } from "./TabLabels";
