+++ src/browser/features/RightSidebar/Tabs/registry.ts
@@ -80,7 +80,7 @@ export interface TabConfig {
 
 /** Static tab configurations (non-terminal tabs) */
 export const TAB_CONFIGS: Record<
-  "costs" | "review" | "explorer" | "desktop" | "browser" | "output" | "debug",
+  "costs" | "review" | "desktop" | "browser" | "output" | "debug",
   TabConfig
 > = {
   costs: {
@@ -89,10 +89,6 @@ export const TAB_CONFIGS: Record<
   },
   review: {
     name: "Review",
-    contentClassName: "overflow-y-auto p-0",
-  },
-  explorer: {
-    name: "Explorer",
     contentClassName: "overflow-y-auto p-0",
   },
   desktop: {
@@ -123,29 +119,17 @@ export const TERMINAL_TAB_CONFIG: TabConfig = {
   keepAlive: true,
 };
 
-/** File viewer tab configuration */
-export const FILE_TAB_CONFIG: TabConfig = {
-  name: "File",
-  contentClassName: "overflow-auto p-0",
-  keepAlive: false, // No need to keep rendered when hidden
-};
-
 /** Get config for a tab type */
 export function getTabConfig(tab: TabType): TabConfig {
   if (
     tab === "costs" ||
     tab === "review" ||
-    tab === "explorer" ||
     tab === "desktop" ||
     tab === "browser" ||
     tab === "output" ||
     tab === "debug"
   ) {
     return TAB_CONFIGS[tab];
-  }
-  // File tabs
-  if (tab.startsWith("file:")) {
-    return FILE_TAB_CONFIG;
   }
   // All terminal tabs (including "terminal" placeholder)
   return TERMINAL_TAB_CONFIG;
