+++ src/browser/utils/rightSidebarLayout.test.ts
@@ -5,10 +5,8 @@ import {
   closeSplit,
   dockTabToEdge,
   getDefaultRightSidebarLayoutState,
-  isRightSidebarLayoutState,
   moveTabToTabset,
   parseRightSidebarLayoutState,
-  removeTabEverywhere,
   reorderTabInTabset,
   selectTabInFocusedTabset,
   splitFocusedTabset,
@@ -142,19 +140,19 @@ test("moveTabToTabset removes empty source tabset", () => {
 });
 
 test("reorderTabInTabset reorders tabs within a tabset", () => {
-  // Default layout has ["costs", "review", "explorer"]; reorder costs from 0 to 1
+  // Default layout has ["costs", "review"]; reorder costs from 0 to 1
   const s0 = getDefaultRightSidebarLayoutState("costs");
   const s1 = reorderTabInTabset(s0, "tabset-1", 0, 1);
 
   expect(s1.root.type).toBe("tabset");
   if (s1.root.type !== "tabset") throw new Error("expected tabset");
 
-  expect(s1.root.tabs).toEqual(["review", "costs", "explorer"]);
+  expect(s1.root.tabs).toEqual(["review", "costs"]);
   expect(s1.root.activeTab).toBe("costs");
 });
 
 test("dockTabToEdge splits a tabset and moves the dragged tab into the new pane", () => {
-  // Default layout has ["costs", "review", "explorer"]; drag review into a bottom split
+  // Default layout has ["costs", "review"]; drag review into a bottom split
   const s0 = getDefaultRightSidebarLayoutState("costs");
 
   const s1 = dockTabToEdge(s0, "review", "tabset-1", "tabset-1", "bottom");
@@ -263,149 +261,8 @@ test("closeSplit keeps the specified child", () => {
   expect(s2.root.tabs).toEqual(["review"]);
 });
 
-// --- Parent-tab tracking tests ---
-
-test("removeTabEverywhere activates parent tab when parentTab entry exists", () => {
-  const s: RightSidebarLayoutState = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: {
-      type: "tabset",
-      id: "tabset-1",
-      tabs: ["costs", "review", "file:src/foo.ts"],
-      activeTab: "file:src/foo.ts",
-    },
-    parentTab: { "file:src/foo.ts": "review" },
-  };
-
-  const result = removeTabEverywhere(s, "file:src/foo.ts");
-  if (result.root.type !== "tabset") throw new Error("expected tabset");
-  // Should activate "review" (the parent) instead of positional adjacency
-  expect(result.root.activeTab).toBe("review");
-  expect(result.root.tabs).toEqual(["costs", "review"]);
-  // parentTab entry should be cleaned up
-  expect(result.parentTab).toBeUndefined();
-});
-
-test("removeTabEverywhere falls back to positional adjacency when parent is not in same tabset", () => {
-  const s: RightSidebarLayoutState = {
-    version: 1,
-    nextId: 3,
-    focusedTabsetId: "tabset-1",
-    root: {
-      type: "split",
-      id: "split-1",
-      direction: "vertical",
-      sizes: [50, 50],
-      children: [
-        {
-          type: "tabset",
-          id: "tabset-1",
-          tabs: ["costs", "file:src/foo.ts"],
-          activeTab: "file:src/foo.ts",
-        },
-        { type: "tabset", id: "tabset-2", tabs: ["review"], activeTab: "review" },
-      ],
-    },
-    // Parent "review" is in tabset-2, but the file tab is in tabset-1
-    parentTab: { "file:src/foo.ts": "review" },
-  };
-
-  const result = removeTabEverywhere(s, "file:src/foo.ts");
-  // Split remains because tabset-1 still has "costs"
-  expect(result.root.type).toBe("split");
-  if (result.root.type !== "split") throw new Error("expected split");
-  const left = result.root.children[0];
-  if (left.type !== "tabset") throw new Error("expected tabset");
-  // Since "review" is not in tabset-1, fallback to positional adjacency → "costs"
-  expect(left.activeTab).toBe("costs");
-  expect(left.tabs).toEqual(["costs"]);
-});
-
-test("removeTabEverywhere cleans up parentTab entry for the removed tab", () => {
-  const s: RightSidebarLayoutState = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: {
-      type: "tabset",
-      id: "tabset-1",
-      tabs: ["costs", "review", "file:src/a.ts", "file:src/b.ts"],
-      activeTab: "file:src/a.ts",
-    },
-    parentTab: {
-      "file:src/a.ts": "review",
-      "file:src/b.ts": "costs",
-    },
-  };
-
-  const result = removeTabEverywhere(s, "file:src/a.ts");
-  // Only the entry for b.ts should remain
-  expect(result.parentTab).toEqual({ "file:src/b.ts": "costs" });
-});
-
-test("removeTabEverywhere cleans up parentTab entries pointing to a removed parent tab", () => {
-  const s: RightSidebarLayoutState = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: {
-      type: "tabset",
-      id: "tabset-1",
-      tabs: ["costs", "review", "file:src/a.ts", "file:src/b.ts"],
-      activeTab: "costs",
-    },
-    parentTab: {
-      "file:src/a.ts": "file:src/b.ts",
-      "file:src/b.ts": "review",
-    },
-  };
-
-  // Remove file:src/b.ts — the entry for file:src/a.ts should also be cleaned
-  // because its parent (file:src/b.ts) was the tab that was removed.
-  const result = removeTabEverywhere(s, "file:src/b.ts");
-  expect(result.parentTab).toBeUndefined();
-});
-
-test("persisted layouts without parentTab field still validate correctly (backward compat)", () => {
-  const raw = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: { type: "tabset", id: "tabset-1", tabs: ["costs"], activeTab: "costs" },
-  };
-  expect(isRightSidebarLayoutState(raw)).toBe(true);
-});
-
-test("persisted layouts with valid parentTab field validate correctly", () => {
-  const raw = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: {
-      type: "tabset",
-      id: "tabset-1",
-      tabs: ["costs", "file:src/foo.ts"],
-      activeTab: "costs",
-    },
-    parentTab: { "file:src/foo.ts": "costs" },
-  };
-  expect(isRightSidebarLayoutState(raw)).toBe(true);
-});
-
-test("persisted layouts with invalid parentTab field are rejected", () => {
-  const raw = {
-    version: 1,
-    nextId: 2,
-    focusedTabsetId: "tabset-1",
-    root: { type: "tabset", id: "tabset-1", tabs: ["costs"], activeTab: "costs" },
-    parentTab: { "file:src/foo.ts": 42 }, // value is not a string
-  };
-  expect(isRightSidebarLayoutState(raw)).toBe(false);
-});
-test("parseRightSidebarLayoutState strips legacy 'stats' tabs from persisted layouts", () => {
-  // Simulate a persisted layout that still contains the old "stats" tab
+test("parseRightSidebarLayoutState strips removed static tabs from persisted layouts", () => {
+  // Simulate a persisted layout that still contains removed tabs
   const raw = {
     version: 1,
     nextId: 2,
@@ -424,8 +281,8 @@ test("parseRightSidebarLayoutState strips legacy 'stats' tabs from persisted lay
   expect(result.root.type).toBe("tabset");
   if (result.root.type !== "tabset") throw new Error("expected tabset");
 
-  // "stats" should be stripped, then always-visible tabs are re-injected.
-  expect(result.root.tabs).toEqual(["costs", "review", "explorer"]);
+  // Removed tabs should be stripped without being re-injected.
+  expect(result.root.tabs).toEqual(["costs", "review"]);
   expect(result.root.activeTab).toBe("costs");
 });
 
@@ -448,7 +305,7 @@ test("parseRightSidebarLayoutState falls back activeTab when stats was active", 
   if (result.root.type !== "tabset") throw new Error("expected tabset");
 
   // "stats" stripped; activeTab should fall back to the first remaining tab.
-  expect(result.root.tabs).toEqual(["costs", "review", "explorer"]);
+  expect(result.root.tabs).toEqual(["costs", "review"]);
   expect(result.root.activeTab).toBe("costs");
 });
 test("parseRightSidebarLayoutState maps stats activeTab to costs even when reordered", () => {
@@ -471,11 +328,11 @@ test("parseRightSidebarLayoutState maps stats activeTab to costs even when reord
   if (result.root.type !== "tabset") throw new Error("expected tabset");
 
   // "stats" stripped; activeTab should map to "costs" (semantic replacement), not "review".
-  expect(result.root.tabs).toEqual(["review", "costs", "explorer"]);
+  expect(result.root.tabs).toEqual(["review", "costs"]);
   expect(result.root.activeTab).toBe("costs");
 });
 
-test("parseRightSidebarLayoutState handles split layouts with legacy stats", () => {
+test("parseRightSidebarLayoutState handles split layouts with removed tabs", () => {
   const raw = {
     version: 1,
     nextId: 3,
@@ -504,7 +361,7 @@ test("parseRightSidebarLayoutState handles split layouts with legacy stats", () 
 
   const result = parseRightSidebarLayoutState(raw, "costs");
 
-  // Both tabsets should have stats stripped
+  // Both tabsets should have removed tabs stripped
   expect(result.root.type).toBe("split");
   if (result.root.type !== "split") throw new Error("expected split");
 
@@ -516,7 +373,7 @@ test("parseRightSidebarLayoutState handles split layouts with legacy stats", () 
 
   if (left.type !== "tabset" || right.type !== "tabset") throw new Error("expected tabsets");
 
-  expect(left.tabs).toEqual(["costs", "explorer"]);
+  expect(left.tabs).toEqual(["costs"]);
   expect(left.activeTab).toBe("costs");
   expect(right.tabs).toEqual(["review"]);
   expect(right.activeTab).toBe("review");
